requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
asyncio==3.4.3
openai~=1.9.0
requests~=2.31.0